Write a program that builds a dictionary of terms based on a given collection of text files.
Text files can be submitted in any format.
The size of text files is at least 150 K.
The number of text files is at least 10.
Glossary of terms to save to disk.
Estimate the size of the collection, the total number of words in the collection and the size of the dictionary.
Justify the data structure
Evaluate the complexity of the algorithm
Try several formats of saving the dictionary (dictionary serialization, saving to a text file, etc.) and compare the results.